requirements.txt
alembic==1.0.5
amqp==2.4.2
aniso8601==4.0.1
appnope==0.1.0
arrow==0.12.1
asn1crypto==0.24.0
attrs==19.1.0
Babel==2.6.0
bcrypt==3.1.6
billiard==3.5.0.5
blinker==1.4
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
bz2file==0.98
celery==4.1.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
decorator==4.3.0
docutils==0.14
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Caching==1.4.0
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Principal==0.4.0
Flask-Redis==0.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gensim==3.7.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.8.1
idna==2.8
idna-ssl==1.1.0
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
jieba==0.39
Jinja2==2.10
jmespath==0.9.4
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.13.6
multidict==4.5.2
mysql==0.0.2
mysqlclient==1.4.2.post1
numpy==1.16.2
paramiko==2.4.2
parso==0.3.1
passlib==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
PyYAML==5.1
redis==2.10.5
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
simplegeneric==0.8.1
six==1.12.0
smart-open==1.8.1
speaklater==1.3
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.33.11
thrift==0.11.0
traitlets==4.3.2
typing-extensions==3.7.2
urllib3==1.24.1
vine==1.2.0
wcwidth==0.1.7
webargs==5.1.1
Werkzeug==0.14.1
WTForms==2.2.1
yarl==1.2.0
